+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - `GetMetadata` now returns characters instead of factors
 - `DtDimension` no longer allows missing id
+- `GetHistory` now returns characters instead of factors
 
 
 # CHANGES IN faosws VERSION 0.9.0